src/services/requirements.txt
boto3==1.39.3
Flask==3.1.1
Flask_Cors==4.0.1
langchain_openai==0.3.27
numpy==2.3.1
pandas==2.3.0
python-dotenv==1.1.1
Requests==2.32.4
scikit_learn==1.7.0
statsmodels==0.14.4
